refactor: ChessResources from UI's screens.chess.game -> Domain's model.game.board.
